refactor(plots): split 1404-line _plots/_data.py into a subpackage
The monolithic _plots/_data.py was ~1.65x the next-largest source module.
Decompose it into a _plots/_data/ package, one module per plot family, with
the DataPlots class recombined from mixins:

- _styling.py     shared layout/colour helpers
- _dashboard.py   performance-dashboard figure builder
- _cumulative.py  returns / compare / log_returns / earnings
- _periodic.py    daily / yearly / monthly bars + monthly heatmap
- _distribution.py histogram / distribution
- _montecarlo.py  montecarlo / montecarlo_distribution
- _drawdown.py    drawdown / drawdowns_periods
- _rolling.py     rolling sharpe / sortino / volatility / beta
- _core.py        DataPlots facade (mixins + snapshot)

Public API is unchanged: `from ._data import DataPlots` and every method
signature are identical. __slots__ is preserved (mixins use empty slots,
the facade owns ("_data",)). The lone internal consumer, _plots/_portfolio.py,
now imports _apply_base_layout from ._data._styling.

No source file now exceeds ~256 lines. All gates green: ruff, ty + mypy
--strict, interrogate 100%, 1210 tests pass at 100% line+branch coverage.
