+++ src/jquantstats/_plots/_data/__init__.py
@@ -1,0 +1,7 @@
+"""Plotting utilities for financial returns data."""
+
+from __future__ import annotations
+
+from ._core import DataPlots
+
+__all__ = ["DataPlots"]
